+++ datacontract/data_contract.py
@@ -41,7 +41,7 @@ class DataContract:
         self,
         data_contract_file: str = None,
         data_contract_str: str = None,
-        data_contract: DataContractSpecification = None,
+        data_contract: DataContractSpecification | OpenDataContractStandard = None,
         schema_location: str = None,
         server: str = None,
         publish_url: str = None,
@@ -250,8 +250,8 @@ class DataContract:
             inline_quality=self._inline_quality,
         )
 
-    def export(self, export_format: ExportFormat, model: str = "all", sql_server_type: str = "auto", **kwargs) -> str:
-        if export_format == ExportFormat.html or export_format == ExportFormat.mermaid:
+    def export(self, export_format: ExportFormat, model: str = "all", spec: Spec = Spec.datacontract_specification, sql_server_type: str = "auto", **kwargs) -> str:
+        if export_format == ExportFormat.html or export_format == ExportFormat.mermaid or spec == Spec.odcs:
             data_contract = resolve.resolve_data_contract_v2(
                 self._data_contract_file,
                 self._data_contract_str,