+++ datacontract/export/custom_converter.py
@@ -1,8 +1,9 @@
 from pathlib import Path
 
 from jinja2 import Environment, FileSystemLoader
+from open_data_contract_standard.model import OpenDataContractStandard
 
-from datacontract.export.exporter import Exporter
+from datacontract.export.exporter import Exporter, Spec
 from datacontract.model.data_contract_specification import (
     DataContractSpecification,
     Model,
@@ -14,24 +15,27 @@ class CustomExporter(Exporter):
 
     def export(
         self,
-        data_contract: DataContractSpecification,
+        data_contract: DataContractSpecification | OpenDataContractStandard,
         model: Model,
         server: str,
         sql_server_type: str,
         export_args: dict,
+        spec: Spec = Spec.datacontract_specification
     ) -> str:
         """Exports a data contract to custom format with Jinja."""
         template = export_args.get("template")
         if template is None:
             raise RuntimeError("Export to custom requires template argument.")
 
-        return to_custom(data_contract, template)
+        return to_custom(data_contract, template, spec)
 
 
-def to_custom(data_contract: DataContractSpecification, template_path: Path) -> str:
-    template = get_template(template_path)
-    rendered_sql = template.render(data_contract=data_contract)
-    return rendered_sql
+def to_custom(data_contract: DataContractSpecification | OpenDataContractStandard, template_path: str, spec: Spec = Spec.datacontract_specification) -> str:
+    template = get_template(Path(template_path))
+
+    # Add the specification value to the rendering context
+    rendered_output = template.render(data_contract=data_contract, spec=spec.value)
+    return rendered_output
 
 
 def get_template(path: Path):